WIP: T02 -- relocate is_event_type to events, add utils->events RULE
